Rename `TypeMappings` to `StandardMappings`
