+++ rdbms/src/test/java/io/spine/server/storage/jdbc/projection/JdbcProjectionStorageShould.java
@@ -35,7 +35,7 @@ import io.spine.test.Tests;
 import io.spine.test.storage.ProjectId;
 import org.junit.Test;
 
-import static io.spine.server.storage.jdbc.TypeMappings.mySql;
+import static io.spine.server.storage.jdbc.StandardMappings.mySql;
 import static io.spine.test.Tests.nullRef;
 import static org.junit.Assert.assertNotNull;
